+++ src/compas/robots/resources/github.py
@@ -8,10 +8,9 @@ from compas.robots.resources.basic import _mesh_import
 __all__ = ['GithubPackageMeshLoader']
 
 try:
-    from urllib.request import urlopen, urlretrieve
+    from urllib.request import urlopen
 except ImportError:
     from urllib2 import urlopen
-    from urllib import urlretrieve
 
 
 class GithubPackageMeshLoader(AbstractMeshLoader):
@@ -27,16 +26,21 @@ class GithubPackageMeshLoader(AbstractMeshLoader):
         and additional assets, e.g. 'abb_irb4400_support'
     branch : str
         Branch name, defaults to ``main``.
+    relative_path : str
+        Relative path of the support package within the repository.
+        If the repository itself is the support package, set
+        ``relative_path`` to ``'.'``.  Defaults to ``support_package``
     """
 
     HOST = 'https://raw.githubusercontent.com'
 
-    def __init__(self, repository, support_package, branch='main'):
+    def __init__(self, repository, support_package, branch='main', relative_path=None):
         super(GithubPackageMeshLoader, self).__init__()
         self.repository = repository
         self.support_package = support_package
         self.branch = branch
         self.schema_prefix = 'package://' + self.support_package + '/'
+        self.relative_path = support_package if relative_path is None else relative_path
 
     def build_url(self, file):
         """Returns the corresponding url of the file.
@@ -52,11 +56,15 @@ class GithubPackageMeshLoader(AbstractMeshLoader):
         str
             The file's url.
         """
-        return '{}/{}/{}/{}/{}'.format(GithubPackageMeshLoader.HOST,
-                                       self.repository,
-                                       self.branch,
-                                       self.support_package,
-                                       file)
+        url_components = [
+            GithubPackageMeshLoader.HOST,
+            self.repository,
+            self.branch,
+            file
+        ]
+        if self.relative_path != '.':
+            url_components.insert(3, self.relative_path)
+        return '/'.join(url_components)
 
     def load_urdf(self, file):
         """Load a URDF file from a Github support package repository.
@@ -102,8 +110,4 @@ class GithubPackageMeshLoader(AbstractMeshLoader):
         _prefix, path = url.split(self.schema_prefix)
         url = self.build_url(path)
 
-        # TODO: As soon as compas.files adds support
-        # for file-like objects, we could skip
-        # storing a temp file for these urls
-        tempfile, _ = urlretrieve(url)
-        return _mesh_import(url, tempfile)
+        return _mesh_import(url, url)