+++ CHANGELOG.md
@@ -74,6 +74,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Added Grasshopper components for JSON serialization.
 * Added `compas_rhino.utilities.set_object_attributes`.
 * Added `from_jsonstring` and `to_jsonstring`.
+* Added Grasshopper component documentation.
 
 ### Changed