+++ CHANGELOG.md
@@ -21,6 +21,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 * Allow str or int as joint type in `compas.robots.Joint` constructor.
 * `compas_ghpython.artists.FrameArtist.draw` now draws a Rhino Plane.
+* Changed directory where ghuser components are installed.
+* Added ghuser components directory to those removed by the `clean` task.
 
 ### Removed