+++ CHANGELOG.md
@@ -24,6 +24,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Fixed bugs in `compas.geometry.bestfit_circle_numpy`.
 * Changed directory where ghuser components are installed.
 * Added ghuser components directory to those removed by the `clean` task.
+* Clean up the ghuser directory before building ghuser components.
 
 ### Removed