+++ CHANGELOG.md
@@ -6,13 +6,23 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## Unreleased
+
+### Added
+
+### Changed
+* `compas.robots.Axis` is now normalized upon initialization.
+
+### Removed
+
+
 ## [1.7.1] 2021-06-14
 
 ### Added
 
 ### Changed
 
-* Fixed bundling of ghuser components
+* Fixed bundling of ghuser components.
 
 ### Removed
 
@@ -26,7 +36,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Added pluggable function `trimesh_principal_curvature` in `compas_rhino`.
 * Added `copy` and `deepcopy` functionality to `compas.robots.Configuration`.
 * Added grasshopper component for drawing a frame.
-* Added `draw_origin` and `draw_axes`
+* Added `draw_origin` and `draw_axes`.
 
 ### Changed